Split some shard map and coordinator functionality into source files. Address some clang-tidy feedback
