+++ src/io/messages.hpp
@@ -23,7 +23,6 @@ namespace memgraph::io::messages {
 using memgraph::coordinator::CoordinatorReadRequests;
 using memgraph::coordinator::CoordinatorWriteRequests;
 using memgraph::coordinator::CoordinatorWriteResponses;
-using memgraph::coordinator::HeartbeatResponse;
 
 // TODO(everbody) change these to the real shard messages
 using memgraph::io::rsm::StorageReadRequest;
@@ -32,7 +31,6 @@ using memgraph::io::rsm::StorageWriteRequest;
 using memgraph::io::rsm::AppendRequest;
 using memgraph::io::rsm::AppendResponse;
 using memgraph::io::rsm::ReadRequest;
-using memgraph::io::rsm::ReadResponse;
 using memgraph::io::rsm::VoteRequest;
 using memgraph::io::rsm::VoteResponse;
 using memgraph::io::rsm::WriteRequest;
